test2

+++ build.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
-rm -rf ./dist
+git pull
+npm install
+if [ -d ./dist ];then
+	echo "removing dist dir"
+	rm -rf ./dist
+fi
 npm install
 npm run build:prod
+echo 'success'
+
